+++ src/simplify.jl
@@ -240,6 +240,10 @@ This function checks all sum constraints which are completely inside of the alld
 function simplify_all_different_inner_equal_to(com, constraint::AllDifferentConstraint)
     @assert length(constraint.indices) == length(constraint.pvals)
     added_constraint_idxs = Int[]
+    # make sure that we don't use an index more than once
+    used_indices = falses(length(com.search_space))
+    # we don't try to maximize the sum here so one might be able to use
+    # a different constraint to start with to get a better sum.
     all_diff_sum = sum(constraint.pvals)
     in_sum = 0
     found_possible_constraint = false
@@ -251,11 +255,16 @@ function simplify_all_different_inner_equal_to(com, constraint::AllDifferentCons
         if isa(sub_constraint.fct, SAF) && isa(sub_constraint.set, MOI.EqualTo)
             # the coefficients must be all 1
             if all(t.coefficient == 1 for t in sub_constraint.fct.terms)
-                # compute sum inside all sum constraints
-                found_possible_constraint = true
-                in_sum += sub_constraint.set.value - sub_constraint.fct.constant
-                # for sum which are in alldifferent but not in sum constraints
-                outside_indices = setdiff(outside_indices, sub_constraint.indices)
+                # only use constraint if none of the indices was already used for
+                # computing the inner sum
+                if !any(used_indices[sub_constraint.indices])
+                    used_indices[sub_constraint.indices] .= true
+                    # compute sum inside all sum constraints
+                    found_possible_constraint = true
+                    in_sum += sub_constraint.set.value - sub_constraint.fct.constant
+                    # for sum which are in alldifferent but not in sum constraints
+                    outside_indices = setdiff(outside_indices, sub_constraint.indices)
+                end
             end
         end
     end