+++ test/runtests.jl
@@ -9,14 +9,26 @@ const MOI = MathOptInterface
 const CS = ConstraintSolver
 const MOIU = MOI.Utilities
 
-CSTestOptimizer() = CS.Optimizer(logging = [])
-CSJuMPTestOptimizer() = JuMP.optimizer_with_attributes(CS.Optimizer, "logging" => [])
+function CSTestOptimizer(; branch_strategy=:Auto)
+    CS.Optimizer(logging = [], seed=1, branch_strategy=branch_strategy)
+end
+function CSJuMPTestOptimizer(; branch_strategy=:Auto)
+    JuMP.optimizer_with_attributes(CS.Optimizer,
+        "logging" => [],
+        "seed"=>4,
+        "branch_strategy"=>branch_strategy
+    )
+end
 cbc_optimizer = optimizer_with_attributes(Cbc.Optimizer, "logLevel" => 0)
-CSCbcJuMPTestOptimizer() = JuMP.optimizer_with_attributes(
-    CS.Optimizer,
-    "logging" => [],
-    "lp_optimizer" => cbc_optimizer,
-)
+function CSCbcJuMPTestOptimizer(; branch_strategy=:Auto)
+    JuMP.optimizer_with_attributes(
+        CS.Optimizer,
+        "logging" => [],
+        "lp_optimizer" => cbc_optimizer,
+        "seed" => 2,
+        "branch_strategy" => branch_strategy
+    )
+end
 
 macro test_macro_throws(errortype, m)
     # See https://discourse.julialang.org/t/test-throws-with-macros-after-pr-23533/5878